+++ src/misc.c
@@ -30,7 +30,7 @@ void ppcGenPosition(S_position *p) {
 	fprintf(ccOut, "<%s %s=%d %s=%d %s=%d>%s</%s>\n", 
 			PPC_LC_POSITION, 
 			PPCA_LINE, p->line, PPCA_COL, p->coll, 
-			PPCA_LEN, strlen(fn), fn, 
+			PPCA_LEN, (int)strlen(fn), fn,
 			PPC_LC_POSITION);
 	//&ppcGenRecord(PPC_FILE, s_fileTab.tab[p->file]->name,"\n");
 	//&ppcGenNumericRecord(PPC_LINE, p->line,"","");
@@ -54,7 +54,7 @@ void ppcGenOffsetPosition(char *fn, int offset) {
 	fprintf(ccOut, "<%s %s=%d %s=%d>%s</%s>\n", 
 			PPC_OFFSET_POSITION, 
 			PPCA_OFFSET, offset, 
-			PPCA_LEN, strlen(fn), fn, 
+			PPCA_LEN, (int)strlen(fn), fn,
 			PPC_OFFSET_POSITION);
 	//&ppcGenRecord(PPC_FILE, m->buffer->name,"\n");
 	//&ppcGenNumericRecord(PPC_OFFSET, m->offset,"","");
@@ -117,7 +117,7 @@ void ppcGenAllCompletionsRecordBegin(int nofocus, int len) {
 void ppcGenRecordWithNumeric(char *kind, char *attr, int val, char *message,char *suff) {
 	ppcIndentOffset();
 	fprintf(ccOut, "<%s %s=%d %s=%d>%s</%s>%s", kind, 
-			attr, val, PPCA_LEN, strlen(message),
+			attr, val, PPCA_LEN, (int)strlen(message),
 			message, kind, suff);
 }
 
@@ -126,7 +126,7 @@ void ppcGenTwoNumericsAndrecord(char *kind, char *attr1, int val1, char *attr2, 
 	fprintf(ccOut, "<%s %s=%d %s=%d %s=%d>%s</%s>%s", kind, 
 			attr1, val1, 
 			attr2, val2, 
-			PPCA_LEN, strlen(message),
+			PPCA_LEN, (int)strlen(message),
 			message, kind, suff);
 }
 
@@ -137,7 +137,7 @@ void ppcGenNumericRecord(char *kind, int val,char *message,char *suff) {
 
 void ppcGenRecord(char *kind, char *message, char *suffix) {
 	ppcIndentOffset();
-	fprintf(ccOut, "<%s %s=%d>%s</%s>%s", kind, PPCA_LEN, strlen(message), message, 
+	fprintf(ccOut, "<%s %s=%d>%s</%s>%s", kind, PPCA_LEN, (int)strlen(message), message,
 			kind, suffix);
 }
 
@@ -151,7 +151,7 @@ void ppcGenTmpBuff() {
 void ppcGenDisplaySelectionRecord(char *message, int messageType, int continuation) {
 	ppcIndentOffset();
 	fprintf(ccOut, "<%s %s=%d %s=%d %s=%d>%s</%s>\n", PPC_DISPLAY_RESOLUTION, 
-			PPCA_LEN, strlen(message), 
+			PPCA_LEN, (int)strlen(message),
 			PPCA_MTYPE, messageType,
 			PPCA_CONTINUE, (continuation==CONTINUATION_ENABLED) ? 1 : 0,
 			message, PPC_DISPLAY_RESOLUTION);
